# swagger.yaml
swagger: "2.0"
info:
  title: Bank Argent API documentation
  description: Contains all available API endpoints in this codebase
  version: "1.0.0"
  termsOfService: "http://swagger.io/terms/"
host: localhost:3001
basePath: /api/v1
schemes:
  - http
paths:
  # User Moduls
  /user/login:
    post:
      tags:
        - User Module
      summary: Login
      description: API for Login
      parameters:
        - in: body
          name: body
          description: Login Payload
          required: true
          schema:
            $ref: "#/definitions/Login"
      produces:
        - application/json
      responses:
        "200":
          description: Login Successfully
          schema:
            $ref: "#/definitions/LoginResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
  /user/signup:
    post:
      tags:
        - User Module
      summary: Signup
      description: API for Signup
      parameters:
        - in: body
          name: body
          description: Signup Payload
          required: true
          schema:
            $ref: "#/definitions/User"
      produces:
        - application/json
      responses:
        "200":
          description: Signup Successfully
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
  /user/profile:
    post:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: User Profile API
      description: API for fetching a user profile
      #produces:
      # - application/json
      responses:
        "200":
          description: User profile retrieved successully
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
    put:
      security:
        - Bearer: []
      tags:
        - User Module
      summary: User Profile API
      description: API for updating a user profile
      parameters:
        - in: body
          name: body
          description: Update user profile attributes
          required: true
          schema:
            $ref: "#/definitions/UserProfile"
      produces:
        - application/json
      responses:
        "200":
          description: User profile retrieved successully
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
  # Account  Moduls
  /account/getAccount:
    get:
      security:
        - Bearer: []
      tags:
        - Account Module
      summary: Get Account Information
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: getAccount retrieved successully
          schema:
            $ref: "#/definitions/Account"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
  # Transaction  Moduls
  /transaction/createTransactions:
    post:
      security:
        - Bearer: []
      tags:
        - Transaction Module
      summary: Create Transactions
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
        - in: body
          name: body
          description: Transactions Payload
          required: true
          schema:
            $ref: "#/definitions/Transaction"
      produces:
        - application/json
      responses:
        "200":
          description: Transactions retrieved successully
          schema:
            $ref: "#/definitions/Transaction"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
  /transaction/getTransactions:
    get:
      security:
        - Bearer: []
      tags:
        - Transaction Module
      summary: Get Account Transactions
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
        - in: path
          name: accountId
          description: ID of the user's account
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Transactions retrieved successully
          schema:
            $ref: "#/definitions/Transaction"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
  /transaction/updateTransactions:
    put:
      security:
        - Bearer: []
      tags:
        - Transaction Module
      summary: Update Transactions
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
        - in: path
          name: transactionId
          description: ID of the transaction
          required: true
          type: string
        - in: body
          name: body
          description: UpdateTransactions Payload
          required: true
          schema:
            $ref: "#/definitions/UpdateTransactions"
      produces:
        - application/json
      responses:
        "200":
          description: Transactions updated successully
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
  /transaction/deleteTransactionsInfo:
    delete:
      security:
        - Bearer: []
      tags:
        - Transaction Module
      summary: Delete Transactions Information from Transaction Record
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
        - in: path
          name: transactionId
          description: ID of the transaction
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Transactions informations deleted successully
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
  User:
    properties:
      email:
        type: string
        description: user email
      password:
        type: string
        description: user password
      firstName:
        type: string
        description: user first name
      lastName:
        type: string
        description: user last name
      userName:
        type: string
        description: user public name
  Login:
    properties:
      email:
        type: string
        description: user email
      password:
        type: string
        description: user password
  ApiResponse:
    type: object
    properties:
      status:
        type: integer
      message:
        type: string
      body:
        type: object
        properties:
          id:
            type: string
          email:
            type: string
  LoginResponse:
    type: object
    properties:
      token:
        type: string
  UserProfile:
    type: object
    properties:
      userName:
        type: string
  Account:
    type: object
    properties:
      id:
        type: string
        description: Unique identifier for the account
      userId:
        type: string
        description: Unique identifier for the user
      accountTitle:
        type: string
        description: Title of the account
      amount:
        type: number
        description: Current balance of the account
      description:
        type: string
        description: Description of the account
  Transaction:
    type: object
    properties:
      id:
        type: string
        description: Unique identifier for the transaction
      accountId:
        type: string
        description: ID of the account associated with the transaction
      date:
        type: string
        format: date
        description: Date of the transaction
      description:
        type: string
        description: Description of the transaction
      amount:
        type: number
        description: Amount of the transaction
      balance:
        type: number
        description: Current balance of the account
      transactionType:
        type: string
        description: Type of the transaction
      category:
        type: string
        description: Category of the transaction
      note:
        type: string
        description: Note associated with the transaction
  UpdateTransactions:
    type: object
    properties:
      note:
        type: string
        description: Updated note for the transaction
        example: supermarket
      category:
        type: string
        description: Category of the transaction
        example: Food